redis configuration
create a spring-redis.xml to integrate redis into our project and redis.properties set the properties values of redis configuration. JedisPoolWriper class is to deal with  redis connection pool object. In spring.xml dao, we add the location of the redis.properties to the property placeholder. JedisUtil is to get the connection pool object and get the jedis objects and two inner classes which are Keys and Strings to manipulate on key-value pairs